refactor: merge single-export repository files into repositories/index.ts
