+++ backend/src/controllers/course/getCourseById.ts
@@ -2,7 +2,7 @@ import type { Request, Response } from "express";
 import { namedUUIDType } from "lib";
 import { z } from "zod";
 import { validate } from "../../middleware/zodValidateRequest.js";
-import { courseRepository } from "../../repositories/courseRepository.js";
+import { courseRepository } from "../../repositories/index.js";
 
 const dataSchema = z.object({
   params: z.object({
