+++ backend/src/controllers/timetable/copyTimetable.ts
@@ -3,8 +3,10 @@ import { type degreeEnum, timetableIDType } from "lib";
 import { z } from "zod";
 import { type Section, Timetable, type User } from "../../entity/entities.js";
 import { validate } from "../../middleware/zodValidateRequest.js";
-import { timetableRepository } from "../../repositories/timetableRepository.js";
-import { userRepository } from "../../repositories/userRepository.js";
+import {
+  timetableRepository,
+  userRepository,
+} from "../../repositories/index.js";
 import timetableJSON from "../../timetable.json" with { type: "json" };
 import sqids, { validSqid } from "../../utils/sqids.js";
