+++ backend/src/controllers/timetable/createTimetable.ts
@@ -1,8 +1,10 @@
 import type { Request, Response } from "express";
 import type { degreeEnum } from "lib";
 import { type Section, Timetable, type User } from "../../entity/entities.js";
-import { timetableRepository } from "../../repositories/timetableRepository.js";
-import { userRepository } from "../../repositories/userRepository.js";
+import {
+  timetableRepository,
+  userRepository,
+} from "../../repositories/index.js";
 import timetableJSON from "../../timetable.json" with { type: "json" };
 import sqids from "../../utils/sqids.js";
