+++ backend/src/controllers/user/getUserDetails.ts
@@ -3,7 +3,7 @@ import { namedUUIDType } from "lib";
 import { z } from "zod";
 import type { User } from "../../entity/entities.js";
 import { validate } from "../../middleware/zodValidateRequest.js";
-import { userRepository } from "../../repositories/userRepository.js";
+import { userRepository } from "../../repositories/index.js";
 import sqids from "../../utils/sqids.js";
 
 const dataSchema = z.object({
