--- backend/src/repositories/announcementRepository.ts
@@ -1,4 +1,0 @@
-import { AppDataSource } from "../db.js";
-import { Announcement } from "../entity/entities.js";
-
-export const announcementRepository = AppDataSource.getRepository(Announcement);
