--- backend/src/repositories/courseRepository.ts
@@ -1,4 +1,0 @@
-import { AppDataSource } from "../db.js";
-import { Course } from "../entity/entities.js";
-
-export const courseRepository = AppDataSource.getRepository(Course);
