--- backend/src/repositories/sectionRepository.ts
@@ -1,4 +1,0 @@
-import { AppDataSource } from "../db.js";
-import { Section } from "../entity/entities.js";
-
-export const sectionRepository = AppDataSource.getRepository(Section);
