--- backend/src/repositories/timetableRepository.ts
@@ -1,4 +1,0 @@
-import { AppDataSource } from "../db.js";
-import { Timetable } from "../entity/entities.js";
-
-export const timetableRepository = AppDataSource.getRepository(Timetable);
