--- backend/src/repositories/userRepository.ts
@@ -1,4 +1,0 @@
-import { AppDataSource } from "../db.js";
-import { User } from "../entity/entities.js";
-
-export const userRepository = AppDataSource.getRepository(User);
